Time: 2965 ms (1.86%), Space: 8.1 MB (5.28%) - LeetHub

+++ 0278-first-bad-version/0278-first-bad-version.cpp
@@ -4,7 +4,7 @@
 class Solution {
 public:
     int firstBadVersion(int n) {
-        long long left = 1;
+        int left = 1;
         long long right = n;
         long long ans=0;
